Kl 23

+++ api/presents/presents.controller.js
@@ -3,13 +3,18 @@ const userModel = require('../users/users.model')
 const ChildrenSchema = require('../children/children.model');
 const {ChildrenModel} = require('../children/children.model');
 const { Types, SchemaType } = require('mongoose');
-const { types } = require('joi');
-const multer = require('multer');
+const { ObjectId } = require('mongoose').Types;
+// const { types } = require('joi');
+const Joi = require('joi');
 
 class PresentsController {
 
 async getAllPresentsChild(req,res){
 try {
+    const session = req.session;
+      if (!session) {
+        return res.status(404).send({ message: "Session was not found" });
+      }
 
     const {userId} = req.params  // принимает в строке запроса _id User
     console.log("userId =", userId);
@@ -36,31 +41,33 @@ try {
 } 
 }
 
+
+
   async addPresent(req, res, next) {
     try {
+      const session = req.session;
+      if (!session) {
+        return res.status(404).send({ message: "Session was not found" });
+      }
       req.child = { id: '5fbe5d5d25fad0371495570f' }; //Заглушка, очікування обьекта req.child з id
       req.body.childId = req.child.id;
       let { childId } = req.body;
+      const findsId = (id) => {return ChildrenModel.find({_id:id})};
+      if(!findsId(childId))return res.status(404).send({ message: 'Not found' });
       const splitpatch = req.files ? req.files.map(e => (e.path)) : ""
-      const imagePath = splitpatch ? `http://localhost:1717/`+`${splitpatch}`.split('\\').slice().join('/') : "";
-      const { title, bal} = req.body;
-
-      await ChildrenModel.findById(childId, async (err, child) => {
-
-        const newPresent = await PresentsModel.create({
-          title,
-          childId,
-          bal,
-          image: imagePath,
-          dateCreated: Date.now(),
-        })
-        
-        child.presents.push(newPresent);
-
-        await child.save();
-        res.status(200).send('Present added');
-      });
-
+      const imagePath = splitpatch ? `http://localhost:1717/`+`${splitpatch}`.split('\\').slice(3).join('/') : "";
+      const { title, reward} = req.body;
+      await PresentsModel.create({
+        title,
+        childId,
+        reward,
+        image: imagePath,
+        dateCreated: Date.now(),
+      })
+      const newPresent = await PresentsModel.find();
+      const presentId = newPresent.map(e=>(e._id));
+      await ChildrenModel.findByIdAndUpdate(childId, {$set: {presents: presentId}},{ upsert:true, returnNewDocument : true });
+      res.status(200).send('Present added');
     } catch (err) {
       next(err);
     }
@@ -68,10 +75,23 @@ try {
 
   async removePresent(req, res, next) {
     try {
+      const session = req.session;
+      if (!session) {
+        return res.status(404).send({ message: "Session was not found" });
+      }
       const { presentId } = req.params;
-      const result = await PresentsModel.findByIdAndDelete(presentId);
-      if (!result) return res.status(404).send({ message: 'Not found' });
-      return res.status(201).send('Present deleted');
+      req.child = { id: '5fbe5d5d25fad0371495570f'}; //Заглушка, очікування обьекта req.child з id
+      req.body.childId = req.child.id;
+      let { childId } = req.body;
+      if (ObjectId.isValid(presentId)) {
+        console.log(presentId)
+        const result = await PresentsModel.findByIdAndDelete(presentId);
+        if (!result) return res.status(404).send({ message: 'Not found' });
+        const newPresent = await PresentsModel.find();
+        const presentId = newPresent.map(e=>(e._id));
+        await ChildrenModel.findByIdAndUpdate(childId, {$set: {presents: presentId}},{ upsert:true, returnNewDocument : true });
+        return res.status(201).send("Present deleted"); 
+      }
     } catch (err) {
       next(err);
     }
@@ -80,11 +100,12 @@ try {
     // test
   }
 
+
   validPresent = (req, res, next) => {
     const validator = Joi.object({
-      title: Joi.string().required(),
+      title: Joi.string(),
       childId: Joi.string().required(),
-      bal: Joi.number().required(),
+      reward: Joi.number(),
       image: Joi.string(),
       dateCreated: Joi.date(),
     });
